display advert of a user with the id in parameter

+++ src/AdvertSiteBundle/Controller/AdvertController.php
@@ -9,6 +9,7 @@
 namespace AdvertSiteBundle\Controller;
 
 use AdvertSiteBundle\Entity\Advert;
+use AdvertSiteBundle\Entity\User;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\Validator\Constraints\DateTime;
@@ -36,23 +37,27 @@ class AdvertController extends Controller
         $advert->setPublicationDate(new \DateTime());
         $advert->setTitle("Third publication");
         $advert->setDescription("My pretty description, Yes I Know , I tried to speak English, but It's not amazing...");
-        $advert->setState("not published");
+        $advert->setState("published");
 
         $em = $this->getDoctrine()->getManager();
         $em->persist($advert);
         $em->flush();
 
-        return $this->redirectToRoute('cc');
+        // @TODO Redirect to Route to Create new Advert ( Form )
+        return $this->redirectToRoute('listAdvert');
     }
 
     /**
-     * @Route ("/listAdvert", name="listAdvert")
+     * @Route ("/listAdvert/user/{id}", name="listAdvert")
+     * @TODO request with the id of the login user id
      */
 
-    public function AdvertListAction(){
+    public function AdvertListAction($id){
+        $user_repository = $this->getDoctrine()->getManager()->getRepository('AdvertSiteBundle:User');
+        $user = $user_repository->findBy(array("id"=> $id));
+
         $repository = $this->getDoctrine()->getManager()->getRepository('AdvertSiteBundle:Advert');
-        //$listAdverts = $repository->findAll();
-        $listAdverts = $repository->findBy(array("user" => "nono"));
+        $listAdverts = $repository->findBy(array("user" => $user[0]->getUsername()));
 
         return $this->render('@AdvertSite/Admin/advert.html.twig', array("adverts"=>$listAdverts));
 
@@ -68,8 +73,10 @@ class AdvertController extends Controller
         $em = $this->getDoctrine()->getManager();
         $repository = $this->getDoctrine()->getManager()->getRepository('AdvertSiteBundle:Advert');
         $advert = $repository->findOneBy(array("id"=>$id));
-        $em->remove($advert);
-        $em->flush();
+        if($advert->getState() == "not published"){
+            $em->remove($advert);
+            $em->flush();
+        }
 
         return $this->redirectToRoute('listAdvert');
     }
